Fix a wrong g_object_set() call on 32 bit platforms
The max-size-time property of GstQueue expects a 64 bit
value. Passing in a 32 bit value on 32 bit platforms would
lead to the following nullptr to get mangled in there leading
to bad errors.

Fixes: QTBUG-97379
Pick-to: 6.2 6.2.1

+++ src/multimedia/platform/gstreamer/common/qgstreamervideooutput.cpp
@@ -161,9 +161,9 @@ void QGstreamerVideoOutput::setIsPreview()
     // also avoids blocking the queue in case we have an encodebin attached to the tee as well
     videoQueue.set("leaky", 2 /*downstream*/);
     videoQueue.set("silent", true);
-    videoQueue.set("max-size-buffers", 1);
-    videoQueue.set("max-size-bytes", 0);
-    videoQueue.set("max-size-time", 0);
+    videoQueue.set("max-size-buffers", uint(1));
+    videoQueue.set("max-size-bytes", uint(0));
+    videoQueue.set("max-size-time", quint64(0));
 }
 
 void QGstreamerVideoOutput::flushSubtitles()
